+++ Black Box Game/InteractionComponent.h
@@ -53,7 +53,7 @@ class InteractionComponent : public BlackBoxEngine::Component
     int m_playerGold = 120;
     int m_playerEnergy = 10;
 
-    uint32_t m_delayedDisplayId = 0;
+    uint64_t m_delayedDisplayId = 0;
     bool m_isChangingLevel = false;
 public:
     InteractionComponent(BlackBoxEngine::Actor* pOwner) : Component(pOwner) { /*EMPTY*/ }
